Separate generation out of TheKnot and move internal testing out of Properties + get rid of a bunch of unnecessary language pragmas
