+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Core.hs
@@ -33,7 +33,7 @@ import Cardano.Ledger.Core (Era, EraRule, eraProtVerLow)
 import qualified Constrained.API as CV2 (HasSpec, Specification, genFromSpec, genFromSpecT)
 import Constrained.GenT (GE (..), GenMode (..))
 import qualified Constrained.GenT as CV1 (runGenT)
-import Constrained.TheKnot (shrinkWithSpec, simplifySpec)
+import Constrained.Generation (shrinkWithSpec, simplifySpec)
 import Control.Monad.Cont (ContT (..))
 import Control.Monad.Trans (MonadTrans (..))
 import Control.State.Transition.Extended (STS (..))
