+++ libs/constrained-generators/src/Constrained/Examples/MapMember.hs
@@ -1,7 +1,3 @@
-{-# LANGUAGE GADTs #-}
-{-# LANGUAGE QuasiQuotes #-}
-{-# LANGUAGE ViewPatterns #-}
-
 module Constrained.Examples.MapMember where
 
 import Constrained.API
