+++ libs/constrained-generators/src/Constrained/Examples/Tree.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeOperators #-}
 
 module Constrained.Examples.Tree where
